Update Template Files (#3)
# Pull Request

## Description

This pull request updates the project name from "Repository Name" to
"Travel Map" across various files to reflect the correct branding. The
changes primarily focus on updating the documentation to align with the
new project name.

### Documentation updates:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L1-R5):
Updated the project title and corresponding table of contents links to
reflect the new name "Travel Map."
*
[`docs/SECURITY.md`](diffhunk://#diff-2605206b7e64be69186fabc8e8fed35676e0015832ff630141aa4aecd73354e0L9-R9):
Updated URLs in the "Report a vulnerability" and "Security Updates"
sections to replace "repository-template" with "travel-map" in the
repository links.
[[1]](diffhunk://#diff-2605206b7e64be69186fabc8e8fed35676e0015832ff630141aa4aecd73354e0L9-R9)
[[2]](diffhunk://#diff-2605206b7e64be69186fabc8e8fed35676e0015832ff630141aa4aecd73354e0L23-R23)