Add places travelled to (#32)
# Pull Request

## Description

This pull request updates the `travel-map.js` file to enhance the
functionality of the world map by introducing a list of selected regions
and improving the map's visual appearance.

### Enhancements to map functionality:
* Added a `places` array containing ISO 3166-2 country codes, grouped by
continent, to define selected regions on the map.
* Updated the `jsVectorMap` configuration to include the
`selectedRegions` property, which highlights the specified regions on
the map.

### Visual improvements:
* Modified the `regionStyle` to include a `selected` style, changing the
fill color of selected regions to green.