Add vite config (#38)
# Pull Request

## Description

This pull request introduces a new Vite configuration file to set up the
project for deployment. The most important change is the addition of the
`vite.config.js` file, which specifies the base URL for the application.

Deployment configuration:

*
[`website/vite.config.js`](diffhunk://#diff-99d4cdadc8c7f3b590672aa6a89b587e337588a9ceb4de1ea08a4aefaf106b91R1-R7):
Created a new Vite configuration file with a `base` property set to
`/travel-map/` to define the base URL for the application.